requirements.txt
docopt==0.6.1
github3.py==0.8.1
sh==1.09
